feat: pill identification functionality, and results screen revisited
I've created PillIdentificationResult, ShapePrediction, and PillMatch classes to provide a clear and type-safe structure for the data received from the API. I've updated the PillIdentificationResultsScreen to use these new data models. I also improved the layout to be more user-friendly, using cards to group related information and a numbered list for the matches. (10)